Moved sensitivity code and fixed up run.py
Moved code for sensitivity study to separate script, sensivity.py. That allowed the simplification of run.py and now the user will just choose whether or not to run the code for the sensitivity studies. Also, now the commands to compile the tex file work and the words are all in the tex file for demonstrating the important stuff in the sensitivity study.