Project for plotting the U, V transport
Added a project script for plotting the U and V transport of the drifters for a simulation. Right now it is specific to the dwh_stream_f project since I can't figure out how to do this generally while still using the run file, but this is the starting point.